Online HILIC-UV-ESI-MS/MS analysis of N-glycan derivatives from yak and bovine milk glycoproteins
In MS, the molecular weight of N-glycans increases by 13 Da with each additional sialic acid residue. Labeling the reducing ends of glycoprotein N-glycans in YM and BM with d0-GP and d5-GP, respectively, increased the molecular weights by 111 Da and 116 Da, respectively. After mixing the samples in equal proportions, the same N-glycans formed a pair of peaks, which differed by 5 Da in the MS spectrum.